backend/pyproject.toml
[build-system]
requires = ["setuptools>=80"]
build-backend = "setuptools.build_meta"

[project]
name = "automatic-data-extractor"
dynamic = ["version"]
description = "ADE backend services and shared infrastructure."
readme = "README.md"
requires-python = ">=3.13,<3.15"
authors = [{ name = "CLAC IT Team" }]
license = { text = "Proprietary" }
urls = { "Repository" = "https://github.com/clac-ca/automatic-data-extractor" }

# Runtime deps — pinned for reproducible builds
# Union of API + worker dependencies.
dependencies = [
  # ade-engine provides the CLI and runtime engine.
  "ade-engine @ git+https://github.com/clac-ca/ade-engine@v1.7.9",

  "fastapi==0.128.0",
  "uvicorn[standard]==0.40.0",
  "sqlalchemy==2.0.46",
  "alembic==1.18.3",
  "azure-core==1.38.0",
  "azure-identity==1.25.1",
  "azure-storage-blob==12.28.0",
  "psycopg[binary]==3.3.2",
  "pydantic==2.12.5",
  "pydantic-settings==2.12.0",
  "python-multipart==0.0.21",
  "orjson==3.11.4",
  "httpx==0.28.1",
  "python-dotenv==1.2.1",
  "pyjwt[crypto]==2.10.1",
  "email-validator==2.3.0",
  "openpyxl==3.1.5",
  "greenlet==3.3.1",
  "fastapi-users[sqlalchemy]==15.0.3",
  "authlib==1.6.6",
  "sse-starlette==3.2.0",
  "pwdlib[argon2]==0.3.0",
  "typer==0.20.0",
  "uv==0.9.28",
]

[project.scripts]
ade = "ade_cli.main:ade_app"
ade-api = "ade_cli.main:ade_api_app"
ade-worker = "ade_cli.main:ade_worker_app"
ade-db = "ade_cli.main:ade_db_app"
ade-storage = "ade_cli.main:ade_storage_app"

[dependency-groups]
dev = [
  "pytest==9.0.2",
  "pytest-asyncio==1.3.0",
  "pytest-cov==7.0.0",
  "asgi-lifespan==2.1.0",
  "mypy==1.19.1",
  "ruff==0.14.14",
  "openapi-spec-validator==0.7.2",
]

# ---- Packaging ----
[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = [
  "ade*",
  "ade_api*",
  "ade_worker*",
  "ade_db*",
  "ade_storage*",
]

[tool.setuptools.dynamic]
version = { attr = "ade.version.__version__" }

# Keep type markers + assets needed at runtime.
[tool.setuptools.package-data]
"ade_api" = [
  "py.typed",
  "openapi.json",
  "web/**/*",
]
"ade_db" = [
  "alembic.ini",
  "migrations/**/*",
]

# ---- Tooling configs ----
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
  "tests/api",
  "tests/worker",
]
asyncio_mode = "auto"
addopts = "-ra"
markers = [
  "unit: fast tests with no external dependencies",
  "integration: tests requiring external services (Postgres, etc.)",
  "e2e: end-to-end tests",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
preview = true

[tool.mypy]
python_version = "3.13"
plugins = ["pydantic.mypy"]
warn_unused_ignores = true
disallow_untyped_defs = true
no_implicit_optional = true
pretty = true

[[tool.mypy.overrides]]
module = "ade_api.tests.*"
disallow_untyped_defs = false

[tool.coverage.run]
branch = true
source = ["ade_api", "ade_worker", "ade_db", "ade_storage"]

[tool.coverage.report]
skip_empty = true
show_missing = true
